sonic-mgmt: fix excessive ifconfig calls
`ifconfig -a` is called for every VM in
a topology within
ansible/roles/vm_set/library/vm_topology.py.
For topologies with a couple hundred VMs,
this can take anywhere from seconds to minutes
for each call to `ifconfig -a`. Multiplying
this across the number of VMs results in
these calls taking up the majority of time
for add_topo to complete.