[action] [PR:18186] [Cisco chassis] Add extra sleep after reboot to fix the flakiness failure (#240)
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Fixes # (issue)
test_acl.py has a high chance to fail on Cisco chassis.
It's a flakiness failure, add extra sleeping time to wait for the routes to be programmed.
### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
 - [ ] Skipped for non-supported platforms
- [x] Test case improvement

### Back port request
- [ ] 202012
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [x] 202405
- [ ] 202411

### Approach
#### What is the motivation for this PR?
test_acl.py has a high chance to fail on Cisco chassis.
It's a flakiness failure, add extra sleeping time to wait for the routes to be programmed.
#### How did you do it?
Add extra sleeping time to wait for the routes to be programmed.
#### How did you verify/test it?
Test on physical testbeds and it stables passes.
#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->