[action] [PR:18018] Snappi: Add the correct queue to select for test_snappi (#255)
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
tests/snappi_tests/test_snappi.py is failing when we use the non-default queue mapping in variables.py. This is because it uses the priority as the queue id.
In this PR we change the code to use the correct queue_id from the variables.py.

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [X] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
 - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [ ] 202012
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [x] 202405
- [x] 202411

### Approach
#### What is the motivation for this PR?
The failure of test_snappi.py for non default queue mapping.

#### How did you do it?
Pls see description.

#### How did you verify/test it?
Ran it in T2 testbed: