[action] [PR:19963] Add an expected log error to PSU test (#617)
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Add an expected log error to PSU test
This test case is testing turning on/off PSU, from the syslog, we can see that while it's turning off one of the PSU, the psud is accessing some the sysfs attributes of the PSU, and read error happened, following are the adjacent three lines of logs, the first line is reported by psud which complaining that some PSU attribute is missing, the second line is from the kernel which indicating deleting some PSU related kernel device, the third line is from hw-mgmt, indicating that received some attribute delete event related to the PSU.

This error message is expected in this test as it's turning off the PSU., The error ignore list for this test should be updated to include this message.

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [x] Testbed and Framework(new/improvement)
- [ ] New Test case
 - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?
Skip expected error in log

#### How did you do it?
Added the regex to the skip

#### How did you verify/test it?
Ran the test with the skip.

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->