Add module check_bgp_ipv6_routes_converged to check routes (#19588) (#621)
Manually cherry-pick PR
https://github.com/sonic-net/sonic-mgmt/pull/19588 to 202412 branch


**Summary**: Add module check_bgp_ipv6_routes_converged to check routes

**Current steps**
1. Generate expected_routes
2. Record start_time
3. Shutdown port
4. Get running_routes by function get_all_bgp_ipv6_routes, ansible
return all routes info from DUT to sonic-mgmt
5. Compare running_routes and expected_routes
6. If not same, repeat step 4-5
7. Record end_time
8. "end_time - start_time" as the bgp route convergence time. If it is
greater than 300s, case fail

**Issue**
In step4, for scale bgp routes setup, the running routes info may exceed
160M. If the network is not fast enough, it may take several minutes for
Ansible to return such a large info and if the routing is inconsistent,
the step need to be repeated, which leads to inaccurate convergence time
of BGP routes.

**Update**
1. Generate expected_routes
2. Call module "check_bgp_ipv6_routes_converged" with "expected_routes"
and "shutdown ports" as parameters and the module is executed on the DUT
3. In the module, it records start_time and shutdown ports
4. Get the running route and compare it with the expected route. If they
are inconsistent, repeat the above steps. If they are consistent, record
the end_time and return the convergence time to sonic-mgmt. All
processes are performed in the DUT, and there is no need to return the
routing information to sonic-mgmt, thus avoiding network delays and
making the calculated convergence time more accurate